+++ src/test/java/com/fasterxml/jackson/databind/ser/jdk/JDKTypeSerializationTest.java
@@ -14,7 +14,6 @@ import org.junit.jupiter.api.Test;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.*;
-import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -107,23 +106,6 @@ public class JDKTypeSerializationTest
     @Test
     public void testInetAddress() throws Exception
     {
-        // Valid IPv4 address
-        InetAddress address = MAPPER.readValue(q("127.0.0.1"), InetAddress.class);
-        assertEquals("127.0.0.1", address.getHostAddress());
-
-        // Valid IPv6 address
-        InetAddress ip6 = MAPPER.readValue(
-                q("2001:db8:85a3:8d3:1319:8a2e:370:7348"), InetAddress.class);
-        assertEquals("2001:db8:85a3:8d3:1319:8a2e:370:7348", ip6.getHostAddress());
-
-        // IPv6 loopback
-        InetAddress loopback6 = MAPPER.readValue(q("::1"), InetAddress.class);
-        assertEquals("0:0:0:0:0:0:0:1", loopback6.getHostAddress());
-    }
-
-    @Test
-    public void testInetAddressSerialization() throws Exception
-    {
         // Default shape: textual form (host part of InetAddress.toString())
         InetAddress input = InetAddress.getByName("127.0.0.1");
         assertEquals(q("127.0.0.1"), MAPPER.writeValueAsString(input));
@@ -143,27 +125,6 @@ public class JDKTypeSerializationTest
         // ... also as a bean property
         assertEquals(String.format("{\"value\":\"%s\"}", named.getHostAddress()),
                 mapper.writeValueAsString(new InetAddressBean(named)));
-    }
-
-    @Test
-    public void testInetAddressNoDNSLookup() throws Exception
-    {
-        // [databind#6058]: should NOT perform DNS lookup — only IP address literals accepted
-        expectInvalidInetAddress("localhost");
-        expectInvalidInetAddress("google.com");
-        // starts off looking like an IP address (edge case)
-        expectInvalidInetAddress("1.2.3.4.example.com");
-        // punycode
-        expectInvalidInetAddress("xn--bcher-kva.example.com");
-    }
-
-    private void expectInvalidInetAddress(String address) throws Exception {
-        try {
-            MAPPER.readValue(q(address), InetAddress.class);
-            fail("Should not pass");
-        } catch (InvalidFormatException e) {
-            verifyException(e, "Not a valid IP address string literal");
-        }
     }
 
     @Test
